Split main.go into config.go, requests.go, handler.go. Also split tests accordingly
