+++ config_test.go
@@ -4,6 +4,8 @@
 // SPDX-License-Identifier: AGPL-3.0-only OR LicenseRef-Element-Commercial
 // Please see LICENSE files in the repository root for full details.
 
+// config_test.go: tests for parseConfig + readKeySecret (config.go).
+
 package main
 
 import (
