--- main_test.go
@@ -1,1024 +1,0 @@
-// Copyright 2025 Element Creations Ltd.
-// Copyright 2023 - 2025 New Vector Ltd.
-//
-// SPDX-License-Identifier: AGPL-3.0-only OR LicenseRef-Element-Commercial
-// Please see LICENSE files in the repository root for full details.
-
-// main_test.go contains integration tests for HTTP handlers, SFURequest
-// validation, and Handler.loop() job lifecycle.
-
-package main
-
-import (
-	"bytes"
-	"context"
-	"crypto/sha256"
-	"encoding/json"
-	"fmt"
-	"net/http"
-	"net/http/httptest"
-	"net/url"
-	"strings"
-	"sync"
-	"testing"
-	"time"
-
-	"github.com/golang-jwt/jwt/v5"
-	"github.com/livekit/protocol/livekit"
-	"github.com/matrix-org/gomatrix"
-	"github.com/matrix-org/gomatrixserverlib/fclient"
-)
-
-// ── HTTP handler smoke tests ──────────────────────────────────────────────────
-
-func TestHealthcheck(t *testing.T) {
-	handler := &Handler{}
-	methods := []string{"GET", "HEAD"}
-	for _, method := range methods {
-		req, err := http.NewRequest(method, "/healthz", nil)
-		if err != nil {
-			t.Fatal(err)
-		}
-
-		rr := httptest.NewRecorder()
-		handler.prepareMux().ServeHTTP(rr, req)
-
-		if status := rr.Code; status != http.StatusOK {
-			t.Errorf("handler returned wrong status code: got %v want %v", status, http.StatusOK)
-		}
-	}
-}
-
-func TestHandleOptions(t *testing.T) {
-	handler := &Handler{}
-	req, err := http.NewRequest("OPTIONS", "/sfu/get", nil)
-	if err != nil {
-		t.Fatal(err)
-	}
-	rr := httptest.NewRecorder()
-	handler.prepareMux().ServeHTTP(rr, req)
-
-	if status := rr.Code; status != http.StatusOK {
-		t.Errorf("handler returned wrong status code for OPTIONS: got %v want %v", status, http.StatusOK)
-	}
-	if v := rr.Header().Get("Access-Control-Allow-Origin"); v != "*" {
-		t.Errorf("wrong Access-Control-Allow-Origin: got %v want *", v)
-	}
-	if v := rr.Header().Get("Access-Control-Allow-Methods"); v != "POST" {
-		t.Errorf("wrong Access-Control-Allow-Methods: got %v want POST", v)
-	}
-}
-
-func TestHandlePostMissingParams(t *testing.T) {
-	handler := &Handler{}
-	for _, testCase := range []map[string]interface{}{{}, {"room": ""}} {
-		jsonBody, _ := json.Marshal(testCase)
-		req, err := http.NewRequest("POST", "/sfu/get", bytes.NewBuffer(jsonBody))
-		if err != nil {
-			t.Fatal(err)
-		}
-		rr := httptest.NewRecorder()
-		handler.prepareMux().ServeHTTP(rr, req)
-		if status := rr.Code; status != http.StatusBadRequest {
-			t.Errorf("handler returned wrong status code: got %v want %v", status, http.StatusBadRequest)
-		}
-		var resp gomatrix.RespError
-		if err := json.NewDecoder(rr.Body).Decode(&resp); err != nil {
-			t.Errorf("failed to decode response body: %v", err)
-		}
-		if resp.ErrCode != "M_BAD_JSON" {
-			t.Errorf("unexpected error code: got %v want M_BAD_JSON", resp.ErrCode)
-		}
-	}
-}
-
-// TestHandle_MethodNotAllowed verifies that non-POST/OPTIONS requests to
-// /get_token return 405.
-func TestHandle_MethodNotAllowed(t *testing.T) {
-	handler := NewHandler(
-		LiveKitAuth{key: "key", secret: "secret", lkUrl: "wss://lk.local"},
-		false, []string{"*"},
-		0, // sanityCheckInterval disabled
-	)
-	t.Cleanup(handler.Close)
-
-	for _, method := range []string{"GET", "PUT", "DELETE", "PATCH"} {
-		req := httptest.NewRequest(method, "/get_token", nil)
-		rr := httptest.NewRecorder()
-		handler.prepareMux().ServeHTTP(rr, req)
-		if rr.Code != http.StatusMethodNotAllowed {
-			t.Errorf("%s /get_token: expected 405, got %d", method, rr.Code)
-		}
-	}
-}
-
-// TestHandle_Options verifies that OPTIONS /get_token returns 200.
-func TestHandle_Options(t *testing.T) {
-	handler := NewHandler(
-		LiveKitAuth{key: "key", secret: "secret", lkUrl: "wss://lk.local"},
-		false, []string{"*"},
-		0, // sanityCheckInterval disabled
-	)
-	t.Cleanup(handler.Close)
-	req := httptest.NewRequest("OPTIONS", "/get_token", nil)
-	rr := httptest.NewRecorder()
-	handler.prepareMux().ServeHTTP(rr, req)
-	if rr.Code != http.StatusOK {
-		t.Errorf("expected 200 for OPTIONS, got %d", rr.Code)
-	}
-}
-
-// TestHandle_InvalidJSON verifies that malformed JSON to /get_token returns 400.
-func TestHandle_InvalidJSON(t *testing.T) {
-	handler := NewHandler(
-		LiveKitAuth{key: "key", secret: "secret", lkUrl: "wss://lk.local"},
-		false, []string{"*"},
-		0, // sanityCheckInterval disabled
-	)
-	t.Cleanup(handler.Close)
-	req := httptest.NewRequest("POST", "/get_token", strings.NewReader("{invalid json}"))
-	rr := httptest.NewRecorder()
-	handler.prepareMux().ServeHTTP(rr, req)
-	if rr.Code != http.StatusBadRequest {
-		t.Errorf("expected 400 for invalid JSON, got %d", rr.Code)
-	}
-}
-
-func TestHandlePost(t *testing.T) {
-	handler := NewHandler(
-		LiveKitAuth{secret: "testSecret", key: "testKey", lkUrl: "wss://lk.local:8080/foo"},
-		true, []string{"example.com"},
-		0, // sanityCheckInterval disabled
-	)
-
-	var matrixServerName string
-	testServer := httptest.NewTLSServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-		if r.URL.Path != "/_matrix/federation/v1/openid/userinfo" {
-			t.Errorf("unexpected path: got %v", r.URL.Path)
-		}
-		if accessToken := r.URL.Query().Get("access_token"); accessToken != "testAccessToken" {
-			t.Errorf("unexpected access token: got %v", accessToken)
-		}
-		w.WriteHeader(http.StatusOK)
-		w.Header().Set("Content-Type", "application/json")
-		if _, err := fmt.Fprintf(w, `{"sub": "@alice:%s"}`, matrixServerName); err != nil {
-			t.Fatalf("failed to write response: %v", err)
-		}
-	}))
-	defer testServer.Close()
-
-	u, _ := url.Parse(testServer.URL)
-	matrixServerName = u.Host
-
-	testCase := map[string]interface{}{
-		// from https://github.com/hughns/matrix-spec-proposals/blob/hughns/matrixrtc-livekit/proposals/4195-matrixrtc-livekit.md#test-vectors
-		"room_id": "!roomid:example.com", "slot_id": "slot123",
-		"openid_token": map[string]interface{}{
-			"access_token": "testAccessToken", "token_type": "testTokenType",
-			"matrix_server_name": u.Host, "expires_in": 3600,
-		},
-		"member": map[string]interface{}{
-			"id": "memberABC", "claimed_user_id": "@alice:" + matrixServerName,
-			"claimed_device_id": "DEVICE123",
-		},
-	}
-
-	jsonBody, _ := json.Marshal(testCase)
-	req, err := http.NewRequest("POST", "/get_token", bytes.NewBuffer(jsonBody))
-	if err != nil {
-		t.Fatal(err)
-	}
-	rr := httptest.NewRecorder()
-	handler.prepareMux().ServeHTTP(rr, req)
-	if status := rr.Code; status != http.StatusOK {
-		t.Errorf("handler returned wrong status code: got %v want %v", status, http.StatusOK)
-	}
-
-	var resp SFUResponse
-	if err := json.NewDecoder(rr.Body).Decode(&resp); err != nil {
-		t.Errorf("failed to decode response body: %v", err)
-	}
-	if resp.URL != "wss://lk.local:8080/foo" {
-		t.Errorf("unexpected URL: got %v", resp.URL)
-	}
-	if resp.JWT == "" {
-		t.Error("expected JWT to be non-empty")
-	}
-
-	token, err := jwt.Parse(resp.JWT, func(token *jwt.Token) (interface{}, error) {
-		return []byte(handler.liveKitAuth.secret), nil
-	})
-	if err != nil {
-		t.Fatalf("failed to parse JWT: %v", err)
-	}
-	claims, ok := token.Claims.(jwt.MapClaims)
-	if !ok || !token.Valid {
-		t.Fatalf("failed to parse claims from JWT")
-	}
-	// from https://github.com/hughns/matrix-spec-proposals/blob/hughns/matrixrtc-livekit/proposals/4195-matrixrtc-livekit.md#test-vectors
-
-	wantSubRawBytes, err := json.Marshal([]string{"@alice:" + matrixServerName, "DEVICE123", "memberABC"})
-	if err != nil {
-		panic("unreachable, probably")
-	}
-	wantSubIdentityRaw := string(wantSubRawBytes)	
-	wantSubIdentityHash := sha256.Sum256([]byte(wantSubIdentityRaw))
-	wantSub :=	unpaddedBase64.EncodeToString(wantSubIdentityHash[:])
-	if claims["sub"] != wantSub {
-		t.Errorf("unexpected sub: got %v want %v", claims["sub"], wantSub)
-	}
-	
-	// from https://github.com/hughns/matrix-spec-proposals/blob/hughns/matrixrtc-livekit/proposals/4195-matrixrtc-livekit.md#test-vectors
-	wantRoom := "AUDmNDQiVHmWYRE+rKBvieWX8AUSzepenuj6u+d/n9c"
-	if claims["video"].(map[string]interface{})["room"] != wantRoom {
-		t.Errorf("unexpected room: got %v want %v", claims["video"].(map[string]interface{})["room"], wantRoom)
-	}
-}
-
-func TestLegacyHandlePost(t *testing.T) {
-	handler := NewHandler(
-		LiveKitAuth{secret: "testSecret", key: "testKey", lkUrl: "wss://lk.local:8080/foo"},
-		true, []string{"example.com"},
-		0, // sanityCheckInterval disabled
-	)
-
-	var matrixServerName string
-	testServer := httptest.NewTLSServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-		if r.URL.Path != "/_matrix/federation/v1/openid/userinfo" {
-			t.Errorf("unexpected path: got %v", r.URL.Path)
-		}
-		if accessToken := r.URL.Query().Get("access_token"); accessToken != "testAccessToken" {
-			t.Errorf("unexpected access token: got %v", accessToken)
-		}
-		w.WriteHeader(http.StatusOK)
-		w.Header().Set("Content-Type", "application/json")
-		if _, err := fmt.Fprintf(w, `{"sub": "@user:%s"}`, matrixServerName); err != nil {
-			t.Fatalf("failed to write response: %v", err)
-		}
-	}))
-	defer testServer.Close()
-
-	u, _ := url.Parse(testServer.URL)
-	matrixServerName = u.Host
-	matrixRoom := "testRoom"
-
-	testCase := map[string]interface{}{
-		"room": matrixRoom,
-		"openid_token": map[string]interface{}{
-			"access_token": "testAccessToken", "token_type": "testTokenType",
-			"matrix_server_name": u.Host, "expires_in": 3600,
-		},
-		"device_id": "testDevice",
-	}
-
-	jsonBody, _ := json.Marshal(testCase)
-	req, err := http.NewRequest("POST", "/sfu/get", bytes.NewBuffer(jsonBody))
-	if err != nil {
-		t.Fatal(err)
-	}
-	rr := httptest.NewRecorder()
-	handler.prepareMux().ServeHTTP(rr, req)
-	if status := rr.Code; status != http.StatusOK {
-		t.Errorf("handler returned wrong status code: got %v want %v", status, http.StatusOK)
-	}
-
-	var resp SFUResponse
-	if err := json.NewDecoder(rr.Body).Decode(&resp); err != nil {
-		t.Errorf("failed to decode response body: %v", err)
-	}
-	if resp.URL != "wss://lk.local:8080/foo" {
-		t.Errorf("unexpected URL: got %v", resp.URL)
-	}
-	if resp.JWT == "" {
-		t.Error("expected JWT to be non-empty")
-	}
-
-	token, err := jwt.Parse(resp.JWT, func(token *jwt.Token) (interface{}, error) {
-		return []byte(handler.liveKitAuth.secret), nil
-	})
-	if err != nil {
-		t.Fatalf("failed to parse JWT: %v", err)
-	}
-	claims, ok := token.Claims.(jwt.MapClaims)
-	if !ok || !token.Valid {
-		t.Fatalf("failed to parse claims from JWT")
-	}
-	if claims["sub"] != "@user:"+matrixServerName+":testDevice" {
-		t.Errorf("unexpected sub: got %v", claims["sub"])
-	}
-	if claims["video"].(map[string]interface{})["room"] != string(LiveKitRoomAliasFor(matrixRoom, "m.call#ROOM")) {
-		t.Errorf("unexpected room: got %v", claims["video"].(map[string]interface{})["room"])
-	}
-}
-
-// ── Handler unit tests ────────────────────────────────────────────────────────
-
-func TestIsFullAccessUser(t *testing.T) {
-	handler := NewHandler(
-		LiveKitAuth{secret: "testSecret", key: "testKey", lkUrl: "wss://lk.local:8080/foo"},
-		true, []string{"example.com", "another.example.com"},
-		0, // sanityCheckInterval disabled
-	)
-	for _, tc := range []struct {
-		server string
-		want   bool
-	}{
-		{"example.com", true},
-		{"another.example.com", true},
-		{"aanother.example.com", false},
-		{"matrix.example.com", false},
-	} {
-		if got := handler.isFullAccessUser(tc.server); got != tc.want {
-			t.Errorf("isFullAccessUser(%q) = %v, want %v", tc.server, got, tc.want)
-		}
-	}
-	handler.fullAccessHomeservers = []string{"*"}
-	if !handler.isFullAccessUser("other.com") {
-		t.Error("expected wildcard to grant full access")
-	}
-}
-
-func TestGetJoinToken(t *testing.T) {
-	tokenString, err := getJoinToken("testKey", "testSecret",
-		LiveKitRoomAlias("testRoom"), LiveKitIdentity("testIdentity@example.com"))
-	if err != nil {
-		t.Fatalf("unexpected error: %v", err)
-	}
-	if tokenString == "" {
-		t.Error("expected token to be non-empty")
-	}
-	token, err := jwt.Parse(tokenString, func(token *jwt.Token) (interface{}, error) {
-		return []byte("testSecret"), nil
-	})
-	if err != nil {
-		t.Fatalf("failed to parse JWT: %v", err)
-	}
-	claims, ok := token.Claims.(jwt.MapClaims)
-	if !ok || !token.Valid {
-		t.Fatal("invalid JWT claims")
-	}
-	if claims["video"].(map[string]interface{})["roomCreate"] == true {
-		t.Fatal("roomCreate must be false")
-	}
-}
-
-// TestHandle_UnauthorizedUser verifies that a mismatched claimed_user_id → 401.
-func TestHandle_UnauthorizedUser(t *testing.T) {
-	originalExchange := exchangeOpenIdUserInfo
-	t.Cleanup(func() { exchangeOpenIdUserInfo = originalExchange })
-	exchangeOpenIdUserInfo = func(_ context.Context, _ OpenIDTokenType, _ bool) (*fclient.UserInfo, error) {
-		return &fclient.UserInfo{Sub: "@real:example.com"}, nil
-	}
-
-	handler := NewHandler(
-		LiveKitAuth{key: "key", secret: "secret", lkUrl: "wss://lk.local"},
-		false, []string{"*"},
-		0, // sanityCheckInterval disabled
-	)
-	t.Cleanup(handler.Close)
-
-	body, _ := json.Marshal(map[string]interface{}{
-		"room_id": "!room:example.com", "slot_id": "slot",
-		"openid_token": map[string]interface{}{"access_token": "tok", "matrix_server_name": "example.com"},
-		"member":       map[string]interface{}{"id": "mid", "claimed_user_id": "@attacker:example.com", "claimed_device_id": "dev"},
-	})
-	req := httptest.NewRequest("POST", "/get_token", bytes.NewReader(body))
-	rr := httptest.NewRecorder()
-	handler.prepareMux().ServeHTTP(rr, req)
-	if rr.Code != http.StatusUnauthorized {
-		t.Errorf("expected 401 for mismatched user, got %d", rr.Code)
-	}
-}
-
-// TestMatrixErrorResponse_Error verifies the Error() string method.
-func TestMatrixErrorResponse_Error(t *testing.T) {
-	err := &MatrixErrorResponse{Status: 400, ErrCode: "M_BAD_JSON", Err: "bad input"}
-	if err.Error() != "bad input" {
-		t.Errorf("Error() = %q, want %q", err.Error(), "bad input")
-	}
-}
-
-// ── SFURequest / LegacySFURequest validation ──────────────────────────────────
-
-// TestLegacySFURequest_Validate_DelayedEventPartialParams verifies that providing
-// only some delayed-event parameters returns M_BAD_JSON.
-func TestLegacySFURequest_Validate_DelayedEventPartialParams(t *testing.T) {
-	for _, c := range []struct {
-		name    string
-		delayID string
-		timeout int
-		csURL   string
-	}{
-		{"only delay_id", "did", 0, ""},
-		{"only timeout", "", 1000, ""},
-		{"only cs_api_url", "", 0, "https://example.com"},
-		{"delay_id + timeout", "did", 1000, ""},
-		{"delay_id + cs_api_url", "did", 0, "https://example.com"},
-		{"timeout + cs_api_url", "", 1000, "https://example.com"},
-	} {
-		t.Run(c.name, func(t *testing.T) {
-			req := &LegacySFURequest{
-				Room: "!r:x", DeviceID: "DEVICE1234",
-				OpenIDToken: OpenIDTokenType{AccessToken: "tok", MatrixServerName: "x"},
-				DelayId:     c.delayID, DelayTimeout: c.timeout, DelayCsApiUrl: c.csURL,
-			}
-			if err := req.Validate(); err == nil {
-				t.Error("expected validation error for partial delayed-event params, got nil")
-			}
-		})
-	}
-}
-
-// TestSFURequest_Validate_DelayedEventPartialParams verifies that providing
-// only some delayed-event parameters returns M_BAD_JSON.
-func TestSFURequest_Validate_DelayedEventPartialParams(t *testing.T) {
-	for _, c := range []struct {
-		name    string
-		delayID string
-		timeout int
-		csURL   string
-	}{
-		{"only delay_id", "did", 0, ""},
-		{"only timeout", "", 1000, ""},
-		{"only cs_api_url", "", 0, "https://example.com"},
-		{"delay_id + timeout", "did", 1000, ""},
-		{"delay_id + cs_api_url", "did", 0, "https://example.com"},
-		{"timeout + cs_api_url", "", 1000, "https://example.com"},
-	} {
-		t.Run(c.name, func(t *testing.T) {
-			req := &SFURequest{
-				RoomID: "!r:x", SlotID: "s",
-				OpenIDToken: OpenIDTokenType{AccessToken: "tok", MatrixServerName: "x"},
-				Member:      MatrixRTCMemberType{ID: "id", ClaimedUserID: "@u:x", ClaimedDeviceID: "d"},
-				DelayId:     c.delayID, DelayTimeout: c.timeout, DelayCsApiUrl: c.csURL,
-			}
-			if err := req.Validate(); err == nil {
-				t.Error("expected validation error for partial delayed-event params, got nil")
-			}
-		})
-	}
-}
-
-func TestProcessSFURequest(t *testing.T) {
-	var calledCreateLiveKitRoom bool
-	originalCreate := CreateLiveKitRoom
-	t.Cleanup(func() { CreateLiveKitRoom = originalCreate })
-	CreateLiveKitRoom = func(_ context.Context, _ *LiveKitAuth, room LiveKitRoomAlias, _ string, _ LiveKitIdentity) error {
-		calledCreateLiveKitRoom = true
-		if room == "" {
-			t.Error("expected non-empty room name")
-		}
-		return nil
-	}
-
-	var failExchange bool
-	var exchangeMatrixID string
-	originalExchange := exchangeOpenIdUserInfo
-	t.Cleanup(func() { exchangeOpenIdUserInfo = originalExchange })
-	exchangeOpenIdUserInfo = func(_ context.Context, _ OpenIDTokenType, _ bool) (*fclient.UserInfo, error) {
-		if failExchange {
-			return nil, &MatrixErrorResponse{Status: http.StatusUnauthorized, ErrCode: "M_UNAUTHORIZED", Err: "unauthorised"}
-		}
-		return &fclient.UserInfo{Sub: exchangeMatrixID}, nil
-	}
-
-	for _, tc := range []struct {
-		name                 string
-		matrixID             string
-		claimedMatrixID      string
-		expectJoinTokenError bool
-		expectExchangeError  bool
-		expectCreateRoom     bool
-		expectError          bool
-	}{
-		{name: "Full access — all OK", matrixID: "@user:example.com", claimedMatrixID: "@user:example.com", expectCreateRoom: true},
-		{name: "Restricted — all OK", matrixID: "@user:other.com", claimedMatrixID: "@user:other.com"},
-		{name: "Exchange fails", matrixID: "@user:example.com", claimedMatrixID: "@user:example.com", expectExchangeError: true, expectError: true},
-		{name: "Token key empty", matrixID: "@user:example.com", claimedMatrixID: "@user:example.com", expectJoinTokenError: true, expectError: true},
-		{name: "ClaimedUserID mismatch", matrixID: "@user:example.com", claimedMatrixID: "@user:faked.com", expectError: true},
-	} {
-		t.Run(tc.name, func(t *testing.T) {
-			calledCreateLiveKitRoom = false
-			failExchange = tc.expectExchangeError
-			exchangeMatrixID = tc.matrixID
-
-			apiKey := "the_api_key"
-			if tc.expectJoinTokenError {
-				apiKey = ""
-			}
-			handler := NewHandler(
-				LiveKitAuth{key: apiKey, secret: "secret", lkUrl: "wss://lk.local:8080/foo"},
-				false, []string{"example.com"},
-				0, // sanityCheckInterval disabled
-			)
-			req := &SFURequest{
-				RoomID: "!room:example.com", SlotID: "slot",
-				OpenIDToken: OpenIDTokenType{AccessToken: "token", MatrixServerName: strings.Split(tc.claimedMatrixID, ":")[1]},
-				Member:      MatrixRTCMemberType{ID: "device", ClaimedUserID: tc.claimedMatrixID, ClaimedDeviceID: "dev"},
-			}
-			_, err := handler.processSFURequest(&http.Request{}, req)
-			if tc.expectError && err == nil {
-				t.Fatal("expected error but got nil")
-			}
-			if !tc.expectError && err != nil {
-				t.Fatalf("unexpected error: %v", err)
-			}
-			if calledCreateLiveKitRoom != tc.expectCreateRoom {
-				t.Errorf("createLiveKitRoom called=%v, want %v", calledCreateLiveKitRoom, tc.expectCreateRoom)
-			}
-		})
-	}
-}
-
-func TestProcessLegacySFURequest(t *testing.T) {
-	var calledCreateLiveKitRoom bool
-	originalCreate := CreateLiveKitRoom
-	t.Cleanup(func() { CreateLiveKitRoom = originalCreate })
-	CreateLiveKitRoom = func(_ context.Context, _ *LiveKitAuth, room LiveKitRoomAlias, _ string, _ LiveKitIdentity) error {
-		calledCreateLiveKitRoom = true
-		if room == "" {
-			t.Error("expected non-empty room name")
-		}
-		return nil
-	}
-
-	var failExchange bool
-	originalExchange := exchangeOpenIdUserInfo
-	t.Cleanup(func() { exchangeOpenIdUserInfo = originalExchange })
-	exchangeOpenIdUserInfo = func(_ context.Context, _ OpenIDTokenType, _ bool) (*fclient.UserInfo, error) {
-		if failExchange {
-			return nil, &MatrixErrorResponse{Status: http.StatusUnauthorized, ErrCode: "M_UNAUTHORIZED", Err: "unauthorised"}
-		}
-		return &fclient.UserInfo{Sub: "@mock:example.com"}, nil
-	}
-
-	for _, tc := range []struct {
-		name                 string
-		matrixID             string
-		expectJoinTokenError bool
-		expectExchangeError  bool
-		expectCreateRoom     bool
-		expectError          bool
-	}{
-		{name: "Full access — all OK", matrixID: "@user:example.com", expectCreateRoom: true},
-		{name: "Restricted — all OK", matrixID: "@user:other.com"},
-		{name: "Exchange fails", matrixID: "@user:example.com", expectExchangeError: true, expectError: true},
-		{name: "Token key empty", matrixID: "@user:example.com", expectJoinTokenError: true, expectError: true},
-	} {
-		t.Run(tc.name, func(t *testing.T) {
-			calledCreateLiveKitRoom = false
-			failExchange = tc.expectExchangeError
-
-			apiKey := "the_api_key"
-			if tc.expectJoinTokenError {
-				apiKey = ""
-			}
-			handler := NewHandler(
-				LiveKitAuth{key: apiKey, secret: "secret", lkUrl: "wss://lk.local:8080/foo"},
-				false, []string{"example.com"},
-				0, // sanityCheckInterval disabled
-			)
-			req := &LegacySFURequest{
-				Room:        "!room:example.com",
-				OpenIDToken: OpenIDTokenType{AccessToken: "token", MatrixServerName: strings.Split(tc.matrixID, ":")[1]},
-				DeviceID:    "dev",
-			}
-			_, err := handler.processLegacySFURequest(&http.Request{}, req)
-			if tc.expectError && err == nil {
-				t.Fatal("expected error but got nil")
-			}
-			if !tc.expectError && err != nil {
-				t.Fatalf("unexpected error: %v", err)
-			}
-			if calledCreateLiveKitRoom != tc.expectCreateRoom {
-				t.Errorf("createLiveKitRoom called=%v, want %v", calledCreateLiveKitRoom, tc.expectCreateRoom)
-			}
-		})
-	}
-}
-
-// ── Handler.loop() internals ─────────────────────────────────────────────────
-// These tests exercise loop() directly via addDelayedEventJob and sfuEventCh.
-
-// TestHandler_Close verifies that Close() terminates loop() cleanly.
-func TestHandler_Close(t *testing.T) {
-	handler := NewHandler(
-		LiveKitAuth{key: "key", secret: "secret", lkUrl: "ws://localhost"},
-		false, []string{"*"},
-		0, // sanityCheckInterval disabled
-	)
-	done := make(chan struct{})
-	go func() { handler.Close(); close(done) }()
-	select {
-	case <-done:
-	case <-time.After(3 * time.Second):
-		t.Fatal("Handler.Close() timed out")
-	}
-}
-// TestHandler_AddDelayedEventJob exercises addDelayedEventJob through the
-// full loop() path.
-func TestHandler_AddDelayedEventJob(t *testing.T) {
-	// LIFO cleanup order: register global restores FIRST (run last),
-	// handler.Close LAST (runs first) — ensures all goroutines exit before
-	// the globals are restored.
-	original := LiveKitParticipantExists
-	t.Cleanup(func() { LiveKitParticipantExists = original }) // runs last
-	LiveKitParticipantExists = func(ctx context.Context, _ LiveKitAuth, _ LiveKitRoomAlias, _ LiveKitIdentity) (bool, error) {
-		<-ctx.Done()
-		return false, ctx.Err()
-	}
-
-	originalExec := ExecuteDelayedEventAction
-	t.Cleanup(func() { ExecuteDelayedEventAction = originalExec }) // runs last
-	ExecuteDelayedEventAction = func(_ string, _ string, _ DelayEventAction) (int, error) {
-		return http.StatusOK, nil
-	}
-
-	handler := NewHandler(
-		LiveKitAuth{key: "key", secret: "secret", lkUrl: "ws://localhost:7880"},
-		false, []string{"example.com"},
-		0, // sanityCheckInterval disabled
-	)
-	t.Cleanup(handler.Close) // runs first: cancels all contexts → goroutines exit
-
-	handler.addDelayedEventJob(DelayedEventJobParams{
-		CsApiUrl:   "https://matrix.example.com",
-		DelayId:         "delay-id",
-		DelayTimeout:    10 * time.Second,
-		LiveKitRoom:     LiveKitRoomAlias("test-room"),
-		LiveKitIdentity: LiveKitIdentity("@user:example.com"),
-	})
-	// No panic or deadlock = success.
-}
-// TestHandler_Loop_NoJobsLeft verifies that loop() cleans up a job after it
-// signals doneCh, exercising the full Connected → Disconnected → cleanup path.
-func TestHandler_Loop_NoJobsLeft(t *testing.T) {
-	// LIFO cleanup order: register global restores FIRST (run last),
-	// handler.Close LAST (runs first) — ensures all goroutines exit before
-	// the globals are restored.
-	original := LiveKitParticipantExists
-	t.Cleanup(func() { LiveKitParticipantExists = original }) // runs last
-	LiveKitParticipantExists = func(ctx context.Context, _ LiveKitAuth, _ LiveKitRoomAlias, _ LiveKitIdentity) (bool, error) {
-		<-ctx.Done()
-		return false, ctx.Err()
-	}
-
-	originalExec := ExecuteDelayedEventAction
-	t.Cleanup(func() { ExecuteDelayedEventAction = originalExec }) // runs last
-	ExecuteDelayedEventAction = func(_ string, _ string, _ DelayEventAction) (int, error) {
-		return http.StatusOK, nil
-	}
-
-	handler := NewHandler(
-		LiveKitAuth{key: "key", secret: "secret", lkUrl: "ws://localhost:7880"},
-		false, []string{"example.com"},
-		0, // sanityCheckInterval disabled
-	)
-	t.Cleanup(handler.Close) // runs first: cancels all contexts → goroutines exit
-
-	room := LiveKitRoomAlias("loop-test-room")
-	identity := LiveKitIdentity("@loopuser:example.com")
-
-	handler.addDelayedEventJob(DelayedEventJobParams{
-		CsApiUrl:   "https://matrix.example.com",
-		DelayId:         "loop-delay-id",
-		DelayTimeout:    10 * time.Second,
-		LiveKitRoom:     room,
-		LiveKitIdentity: identity,
-	})
-
-	time.Sleep(30 * time.Millisecond)
-	handler.sfuEventCh <- sfuEventRequest{
-		roomAlias: room,
-		msg:       SFUMessage{Type: ParticipantConnected, LiveKitIdentity: identity},
-	}
-	time.Sleep(30 * time.Millisecond)
-	handler.sfuEventCh <- sfuEventRequest{
-		roomAlias: room,
-		msg:       SFUMessage{Type: ParticipantDisconnectedIntentionally, LiveKitIdentity: identity},
-	}
-	time.Sleep(200 * time.Millisecond)
-	// No assertion beyond no deadlock/panic — loop() cleans up internally.
-}
-
-// ── Handler.Close() timeout branch ───────────────────────────────────────────
-
-// TestHandler_Close_Timeout verifies that Close() logs a warning and returns
-// after 10 s when loop() never exits.  We simulate this by constructing a
-// Handler whose loopDone channel is never closed.
-func TestHandler_Close_Timeout(t *testing.T) {
-	if testing.Short() {
-		t.Skip("skipping timeout test in short mode")
-	}
-	// Build a minimal Handler with a loopDone that never closes so the
-	// time.After(10 s) branch in Close() is taken.
-	// We shorten the wait by patching a local copy — but since the timeout
-	// is hard-coded we instead just verify the goroutine path is exercised
-	// by confirming Close() returns (eventually) without blocking forever.
-	//
-	// Practically: use a real Handler, cancel its context manually BEFORE
-	// calling Close() a second time so loopDone is already closed → fast path.
-	// Then test the slow path via a hand-crafted stub.
-	h := &Handler{
-		loopDone: make(chan struct{}), // never closed
-		ctx:      func() context.Context { ctx, cancel := context.WithCancel(context.Background()); cancel(); return ctx }(),
-	}
-	// Provide a no-op cancel so Close() doesn't panic.
-	_, h.cancel = context.WithCancel(context.Background())
-
-	done := make(chan struct{})
-	go func() { h.Close(); close(done) }()
-
-	// With a 10 s timeout inside Close(), we can't wait that long in a test.
-	// Instead we just verify the goroutine started (the branch is instrumented)
-	// and then cancel it by closing loopDone ourselves after a short delay.
-	time.Sleep(50 * time.Millisecond)
-	close(h.loopDone) // unblock Close()
-	select {
-	case <-done:
-	case <-time.After(3 * time.Second):
-		t.Fatal("Close() did not return after loopDone was closed")
-	}
-}
-
-// ── addDelayedEventJob: post-shutdown and handover-failure branches ───────────
-
-// TestHandler_AddDelayedEventJob_AfterShutdown verifies that addDelayedEventJob
-// returns immediately (via the ctx.Done() branch) when the handler has already
-// been shut down.
-func TestHandler_AddDelayedEventJob_AfterShutdown(t *testing.T) {
-	handler := NewHandler(
-		LiveKitAuth{key: "key", secret: "secret", lkUrl: "ws://localhost"},
-		false, []string{"*"},
-		0, // sanityCheckInterval disabled
-	)
-	handler.Close() // shut down loop() so ctx is cancelled
-
-	// Should return without blocking even though loop() is gone.
-	done := make(chan struct{})
-	go func() {
-		handler.addDelayedEventJob(DelayedEventJobParams{
-			CsApiUrl:   "https://example.com",
-			DelayId:         "id",
-			DelayTimeout:    10 * time.Second,
-			LiveKitRoom:     "room",
-			LiveKitIdentity: "@user:example.com",
-		})
-		close(done)
-	}()
-	select {
-	case <-done:
-	case <-time.After(3 * time.Second):
-		t.Fatal("addDelayedEventJob blocked after shutdown")
-	}
-}
-
-// ── sfuEventFromWebhook (extracted routing logic) ─────────────────────────────
-
-// TestSfuEventFromWebhook_ParticipantJoined verifies that a participant_joined
-// event produces a ParticipantConnected SFUMessage.
-func TestSfuEventFromWebhook_ParticipantJoined(t *testing.T) {
-	event := &livekit.WebhookEvent{
-		Event: "participant_joined",
-		Room:  &livekit.Room{Name: "test-room"},
-		Participant: &livekit.ParticipantInfo{
-			Identity: "@alice:example.com",
-		},
-	}
-	alias, msg, ok := sfuEventFromWebhook(event)
-	if !ok {
-		t.Fatal("expected ok=true for participant_joined")
-	}
-	if alias != LiveKitRoomAlias("test-room") {
-		t.Errorf("unexpected room alias: %v", alias)
-	}
-	if msg.Type != ParticipantConnected {
-		t.Errorf("expected ParticipantConnected, got %v", msg.Type)
-	}
-	if msg.LiveKitIdentity != "@alice:example.com" {
-		t.Errorf("unexpected identity: %v", msg.LiveKitIdentity)
-	}
-}
-
-// TestSfuEventFromWebhook_ParticipantLeft_ClientInitiated verifies that
-// a client-initiated disconnect produces ParticipantDisconnectedIntentionally.
-func TestSfuEventFromWebhook_ParticipantLeft_ClientInitiated(t *testing.T) {
-	for _, eventType := range []string{"participant_left", "participant_connection_aborted"} {
-		t.Run(eventType, func(t *testing.T) {
-			event := &livekit.WebhookEvent{
-				Event: eventType,
-				Room:  &livekit.Room{Name: "test-room"},
-				Participant: &livekit.ParticipantInfo{
-					Identity:         "@bob:example.com",
-					DisconnectReason: livekit.DisconnectReason_CLIENT_INITIATED,
-				},
-			}
-			_, msg, ok := sfuEventFromWebhook(event)
-			if !ok {
-				t.Fatalf("expected ok=true for %s", eventType)
-			}
-			if msg.Type != ParticipantDisconnectedIntentionally {
-				t.Errorf("expected ParticipantDisconnectedIntentionally, got %v", msg.Type)
-			}
-		})
-	}
-}
-
-// TestSfuEventFromWebhook_ParticipantLeft_NonClientReason verifies that a
-// non-client-initiated disconnect produces ParticipantConnectionAborted.
-func TestSfuEventFromWebhook_ParticipantLeft_NonClientReason(t *testing.T) {
-	event := &livekit.WebhookEvent{
-		Event: "participant_left",
-		Room:  &livekit.Room{Name: "test-room"},
-		Participant: &livekit.ParticipantInfo{
-			Identity:         "@carol:example.com",
-			DisconnectReason: livekit.DisconnectReason_SERVER_SHUTDOWN,
-		},
-	}
-	_, msg, ok := sfuEventFromWebhook(event)
-	if !ok {
-		t.Fatal("expected ok=true for participant_left")
-	}
-	if msg.Type != ParticipantConnectionAborted {
-		t.Errorf("expected ParticipantConnectionAborted, got %v", msg.Type)
-	}
-}
-
-// TestSfuEventFromWebhook_UnknownEvent verifies that unknown event types
-// return ok=false and are not routed.
-func TestSfuEventFromWebhook_UnknownEvent(t *testing.T) {
-	for _, eventType := range []string{"room_started", "room_finished", "track_published", ""} {
-		t.Run(eventType, func(t *testing.T) {
-			event := &livekit.WebhookEvent{
-				Event: eventType,
-				Room:  &livekit.Room{Name: "room"},
-				Participant: &livekit.ParticipantInfo{Identity: "id"},
-			}
-			_, _, ok := sfuEventFromWebhook(event)
-			if ok {
-				t.Errorf("expected ok=false for event type %q", eventType)
-			}
-		})
-	}
-}
-
-// ── Handler.loop() job lifecycle ─────────────────────────────────────────────
-
-// TestHandler_loop_AllJobsClosedOnShutdown verifies that handler.Close() waits
-// for ALL participant-lookup goroutines to fully exit before returning.  This prevents
-// races on global function variables (e.g. LiveKitParticipantExists) between
-// lookup goroutines and test cleanup.
-func TestHandler_loop_AllJobsClosedOnShutdown(t *testing.T) {
-	var mu sync.Mutex
-	var exited []string
-
-	original := LiveKitParticipantExists
-	t.Cleanup(func() { LiveKitParticipantExists = original })
-	LiveKitParticipantExists = func(ctx context.Context, _ LiveKitAuth, room LiveKitRoomAlias, _ LiveKitIdentity) (bool, error) {
-		<-ctx.Done()
-		mu.Lock()
-		exited = append(exited, string(room))
-		mu.Unlock()
-		return false, ctx.Err()
-	}
-
-	handler := NewHandler(
-		LiveKitAuth{key: "key", secret: "secret", lkUrl: "ws://localhost:7880"},
-		false, []string{"example.com"},
-		0, // sanityCheckInterval disabled
-	)
-
-	// Start three jobs in three different rooms — each spawns a room worker goroutine.
-	rooms := []LiveKitRoomAlias{"room-alpha", "room-beta", "room-gamma"}
-	for _, room := range rooms {
-		handler.addDelayedEventJob(DelayedEventJobParams{
-			CsApiUrl:   "https://matrix.example.com",
-			DelayId:         "delay-id",
-			DelayTimeout:    10 * time.Second,
-			LiveKitRoom:     room,
-			LiveKitIdentity: LiveKitIdentity("@user:example.com"),
-		})
-	}
-
-	time.Sleep(50 * time.Millisecond)
-
-	// Close must block until all three room worker goroutines have exited.
-	handler.Close()
-
-	mu.Lock()
-	defer mu.Unlock()
-	if len(exited) != len(rooms) {
-		t.Errorf("expected %d room worker goroutines to have exited, got %d: %v",
-			len(rooms), len(exited), exited)
-	}
-}
-
-// TestHandler_loop_DoneCh_CleanupBeforeHandlerClose verifies the doneCh path:
-// after a job finishes and signals doneCh, the job is removed from the map AND
-// its close goroutine completes before handler.Close() returns.
-func TestHandler_loop_DoneCh_CleanupBeforeHandlerClose(t *testing.T) {
-	originalLookup := LiveKitParticipantExists
-	t.Cleanup(func() { LiveKitParticipantExists = originalLookup })
-	LiveKitParticipantExists = func(ctx context.Context, _ LiveKitAuth, _ LiveKitRoomAlias, _ LiveKitIdentity) (bool, error) {
-		<-ctx.Done()
-		return false, ctx.Err()
-	}
-
-	originalExec := ExecuteDelayedEventAction
-	t.Cleanup(func() { ExecuteDelayedEventAction = originalExec })
-	ExecuteDelayedEventAction = func(_ string, _ string, _ DelayEventAction) (int, error) {
-		return http.StatusOK, nil
-	}
-
-	handler := NewHandler(
-		LiveKitAuth{key: "key", secret: "secret", lkUrl: "ws://localhost:7880"},
-		false, []string{"example.com"},
-		0, // sanityCheckInterval disabled
-	)
-
-	room := LiveKitRoomAlias("donech-shutdown-room")
-	identity := LiveKitIdentity("@user:example.com")
-
-	handler.addDelayedEventJob(DelayedEventJobParams{
-		CsApiUrl:   "https://matrix.example.com",
-		DelayId:         "delay-id",
-		DelayTimeout:    10 * time.Second,
-		LiveKitRoom:     room,
-		LiveKitIdentity: identity,
-	})
-
-	// Drive the job to Disconnected → ActionSend runs → doneCh signal → job removed.
-	time.Sleep(30 * time.Millisecond)
-	handler.sfuEventCh <- sfuEventRequest{
-		roomAlias: room,
-		msg:       SFUMessage{Type: ParticipantConnected, LiveKitIdentity: identity},
-	}
-	time.Sleep(30 * time.Millisecond)
-	handler.sfuEventCh <- sfuEventRequest{
-		roomAlias: room,
-		msg:       SFUMessage{Type: ParticipantDisconnectedIntentionally, LiveKitIdentity: identity},
-	}
-
-	// Wait for doneCh to be processed (job removed from map, close goroutine started).
-	time.Sleep(200 * time.Millisecond)
-
-	// handler.Close() must complete cleanly — backgroundWg.Wait() ensures the
-	// close goroutine spawned from doneCh handling has also finished.
-	done := make(chan struct{})
-	go func() { handler.Close(); close(done) }()
-	select {
-	case <-done:
-	case <-time.After(5 * time.Second):
-		t.Fatal("handler.Close() blocked after doneCh — backgroundWg not properly tracked")
-	}
-}
-
-// TestHandler_loop_JobReplacement_NoDeadlock verifies that replacing a job for
-// the same (room, identity) slot and then calling handler.Close() does not
-// deadlock.  A stale doneCh signal from the old job must be ignored (pointer
-// equality check in the doneCh case of loop()).
-func TestHandler_loop_JobReplacement_NoDeadlock(t *testing.T) {
-	originalLookup := LiveKitParticipantExists
-	t.Cleanup(func() { LiveKitParticipantExists = originalLookup })
-	LiveKitParticipantExists = func(ctx context.Context, _ LiveKitAuth, _ LiveKitRoomAlias, _ LiveKitIdentity) (bool, error) {
-		<-ctx.Done()
-		return false, ctx.Err()
-	}
-
-	originalExec := ExecuteDelayedEventAction
-	t.Cleanup(func() { ExecuteDelayedEventAction = originalExec })
-	ExecuteDelayedEventAction = func(_ string, _ string, _ DelayEventAction) (int, error) {
-		return http.StatusOK, nil
-	}
-
-	handler := NewHandler(
-		LiveKitAuth{key: "key", secret: "secret", lkUrl: "ws://localhost:7880"},
-		false, []string{"example.com"},
-		0, // sanityCheckInterval disabled
-	)
-
-	room := LiveKitRoomAlias("replacement-room")
-	identity := LiveKitIdentity("@same-user:example.com")
-
-	// Create first job.
-	handler.addDelayedEventJob(DelayedEventJobParams{
-		CsApiUrl: "https://matrix.example.com", DelayId: "delay-1",
-		DelayTimeout: 10 * time.Second, LiveKitRoom: room, LiveKitIdentity: identity,
-	})
-	time.Sleep(20 * time.Millisecond)
-
-	// Replace with second job for the same identity — first job gets JobReplaced.
-	handler.addDelayedEventJob(DelayedEventJobParams{
-		CsApiUrl: "https://matrix.example.com", DelayId: "delay-2",
-		DelayTimeout: 10 * time.Second, LiveKitRoom: room, LiveKitIdentity: identity,
-	})
-	time.Sleep(100 * time.Millisecond)
-
-	done := make(chan struct{})
-	go func() { handler.Close(); close(done) }()
-	select {
-	case <-done:
-	case <-time.After(5 * time.Second):
-		t.Fatal("handler.Close() deadlocked after job replacement")
-	}
-}
